requirements.txt
pyodbc>=4.0.39
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# GUI Dependencies
PyQt6>=6.6.0

# Build Dependencies
pyinstaller>=6.0.0
